// frontend/src/components/earth/gis/types.ts
export interface OsmNodeFeature {
  id: string;
  kind: string;
  name?: string;
  lat: number;
  lng: number;
  tags: Record<string, string>;
}